added triggered_screen_blanker_teensy.ino
This is Teensy code that will listen to resonant mirror sync channels and flicker the visual stimulus screen so that it will only by on during the U-turns of the resonant mirrors, when no fluorescence is acquired. Parameters like delays and durations need to be adjusted on the case by case basis.